Improve readPixels() method readability
The readPixels() method optionally requires the pixel
quantity to be read.
The "size" parameter name was misleading, so it has
been changed to "pixels".
Moreover, the method documentation has been clarified,
removing the references to bytes read and returned.
In fact, each pixel requires two bytes, so it was
erroneous to say that 64 bytes are read and returned,
when in reality they are twice that value (128 bytes
in the default case).
This patch does not change the method functionality.

Signed-off-by: Roberto Bonacina <roby.bonacina@tutanota.com>